feat:base-entity interface
